pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "waserver"
description = "Witness Angel Webservices"
version = "0.3"
license = { file = "LICENSE.txt" }
readme = "Readme.rst"
authors = ["Pascal Chambon", "Manon Michelet", "Francinette Atchade"]
homepage = "https://github.com/WitnessAngel/witness-angel-server"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: Unix",
    "Operating System :: MacOS :: MacOS X",
]
include = ["CHANGELOG", "tox.ini", "settings.ini", "docs/*", "docs/api/*", "tests/*.py"]
exclude = ["_build"]

[tool.poetry.dependencies]
python = ">=3.8, <3.13"

# Witness Angel system
wacryptolib = { git = "https://github.com/WitnessAngel/witness-angel-cryptolib.git", branch = "waserver_stable" }

# Django & Bson
Django = ">3,<5"
json-rpc-django = "^1.0"
schema = "^0.7.0"
django_cryptography = "^1.0"
python-decouple = "^3.3"
django-model-utils = "^4.1.1"
djangorestframework = "^3.0"
django-changeset = "^1.1.0"
django-utils-six = "^2.0"
django-compat-patcher = "^0.12"

[tool.poetry.dev-dependencies]

# Test runners and plugins
pytest = ">=7.0.1"
pytest-cov = ">=4.0"
pytest_mock = ">=3.7.0"
pytest-randomly = ">=3.11"
pytest-deadfixtures = ">=2.2.1"
pytest-timeout = ">=2.1.0"
pytest-xdist = ">=3.0.0"
responses = ">=0.18.0"
freezegun = ">=1.1.0"
typeguard = ">=2.13.0"

pytest-django = "^4"
model-bakery = "^1.10.0"
# Linters and formatters
black = {version = ">=22.1.0"}
darglint = ">=1.8.1"
pylint = ">=2.6"
mypy = ">=0.931"
xenon = ">=0.9.0"
safety = ">=1.10"
doc8 = ">=0.10.1"

# Documentation
sphinx = ">=4.4.0"
sphinx-autodoc-typehints = ">=1.17.0"
sphinx_rtd_theme = ">=1.0.0"
toml = ">=0.10.2"

# Frozen executable building
pyinstaller = ">=5"

